requirements.txt
streamlit==1.36.0
ultralytics==8.2.0
opencv-python-headless==4.9.0.80
pillow==10.3.0
numpy==1.26.4